+++ shop/src/main/java/max/shop/domain/User.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 import max.shop.domain.embed.Address;
 import max.shop.domain.time.BaseTimeEntity;
 import max.shop.domain.type.Gender;
-import max.shop.dto.UserRegisterForm;
+import max.shop.dto.request.UserRegisterForm;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
